Added data source for GovUK and Firebase configs
