+++ app/src/main/kotlin/uk/gov/govuk/ui/GovUkApp.kt
@@ -365,6 +365,7 @@ private fun GovUkNavHost(
         )
         analyticsGraph(
             analyticsConsentCompleted = {
+                // TODO GOVUKAPP-3033 activate Firebase here?
                 coroutineScope.launch {
                     appNavigation.onNext(navController)
                 }
