+++ config/src/main/kotlin/uk/gov/govuk/config/data/ConfigRepoImpl.kt
@@ -1,31 +1,22 @@
 package uk.gov.govuk.config.data
 
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import com.google.gson.Gson
+import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.tasks.await
-import uk.gov.govuk.config.SignatureValidator
-import uk.gov.govuk.config.data.remote.ConfigApi
+import uk.gov.govuk.config.FirebaseConfigDataSource
+import uk.gov.govuk.config.GovUkConfigDataSource
 import uk.gov.govuk.config.data.remote.model.ChatUrls
 import uk.gov.govuk.config.data.remote.model.Config
-import uk.gov.govuk.config.data.remote.model.ConfigResponse
 import uk.gov.govuk.config.data.remote.model.EmergencyBanner
 import uk.gov.govuk.config.data.remote.model.UserFeedbackBanner
-import uk.gov.govuk.data.model.Result.DeviceOffline
-import uk.gov.govuk.data.model.Result.Error
-import uk.gov.govuk.data.model.Result.InvalidSignature
+import uk.gov.govuk.data.model.Result
 import uk.gov.govuk.data.model.Result.Success
-import java.net.UnknownHostException
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class ConfigRepoImpl @Inject constructor(
-    private val configApi: ConfigApi,
-    private val gson: Gson,
-    private val signatureValidator: SignatureValidator,
-    private val firebaseRemoteConfig: FirebaseRemoteConfig
+    private val govUkDataSource: GovUkConfigDataSource,
+    private val firebaseDataSource: FirebaseConfigDataSource
 ): ConfigRepo {
 
     companion object {
@@ -35,41 +26,21 @@ class ConfigRepoImpl @Inject constructor(
     private var _config: Config? = null
     private val safeConfig: Config
         get() = checkNotNull(_config) { "You must init config successfully before use!!!" }
-    override suspend fun initConfig(): uk.gov.govuk.data.model.Result<Unit> = coroutineScope {
-        val firebaseJob = launch {
-            try {
-                // Fetch and activate so values are ready immediately
-                firebaseRemoteConfig.fetchAndActivate().await()
-            } catch (e: Exception) {
-                // Log error if needed but don't crash initConfig
-                false
-            }
+    override suspend fun initConfig(): Result<Unit> = coroutineScope {
+        val firebaseConfigDeferred = async {
+            firebaseDataSource.fetchAndActivate()
         }
 
-        try {
-            val response = configApi.getConfig()
-            if (response.isSuccessful) {
-                response.body()?.let {
-                    val signature = response.headers()["x-amz-meta-govuk-sig"] ?: ""
-                    val valid = signatureValidator.isValidSignature(signature, it)
-                    if (!valid) {
-                        return@coroutineScope InvalidSignature()
-                    }
+        val govUkConfigResult = govUkDataSource.fetchConfig()
 
-                    val configResponse = gson.fromJson(it, ConfigResponse::class.java)
+        firebaseConfigDeferred.await()
 
-                    _config = configResponse.config
-
-                    firebaseJob.join()
-                    Success(Unit)
-                } ?: Error()
-            } else {
-                Error()
-            }
-        } catch (_: UnknownHostException) {
-            DeviceOffline()
-        } catch (_: Exception) {
-            Error()
+        if (govUkConfigResult is Success) {
+            _config = govUkConfigResult.value
+            Success(Unit)
+        } else {
+            @Suppress("UNCHECKED_CAST") // we know it's not a success
+            govUkConfigResult as Result<Unit>
         }
     }
 
@@ -116,6 +87,4 @@ class ConfigRepoImpl @Inject constructor(
     override val emergencyBanners: List<EmergencyBanner>?
         get() = safeConfig.emergencyBanners
 
-    val localServicesHeader: String
-        get() = firebaseRemoteConfig.getString("local_services_header")
 }
