+++ feature/home/src/main/kotlin/uk/gov/govuk/home/HomeViewModel.kt
@@ -1,16 +1,13 @@
 package uk.gov.govuk.home
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import uk.gov.govuk.analytics.AnalyticsClient
-import uk.gov.govuk.config.data.ConfigRepo
 import javax.inject.Inject
 
 @HiltViewModel
 internal class HomeViewModel @Inject constructor(
-    private val analyticsClient: AnalyticsClient,
-    private val configRepository: ConfigRepo
+    private val analyticsClient: AnalyticsClient
 ): ViewModel() {
 
     companion object {
@@ -25,7 +22,5 @@ internal class HomeViewModel @Inject constructor(
             screenName = SCREEN_NAME,
             title = TITLE
         )
-
-//        Log.d("HomeViewModel", "local_services_header: ${configRepository.localServicesHeader}")
     }
 }
